dev-requirements.txt
python-dotenv
flask
flask-sqlalchemy
flask-migrate
flask-bcrypt
flask-login
jinjaMarkdown
groq
pytest
pytest-flask
Flask_JWT_Extended
PyJWT
flask-cors
openai
google-generativeai
anthropic